requirements.txt
# Core LLM and orchestration
langchain==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
openai==1.107.2
tenacity==9.1.2
pydantic==2.11.7
python-dotenv==1.1.1
PyYAML==6.0.2

# Document parsing and OCR
PyMuPDF==1.26.4
pytesseract==0.3.13
Pillow==11.3.0
pdf2image==1.17.0

# Retrieval and vector database
chromadb==1.0.20
sentence-transformers==5.1.0
transformers==4.55.4
tiktoken==0.11.0

# Data and evaluation
pandas==2.3.2
numpy==2.3.2
datasets==4.0.0
ragas==0.3.2

# UI
streamlit==1.49.1

# Utilities
tqdm==4.67.1
